+++ android/src/main/java/com/perfektion/fsignalr/FsignalrPlugin.java
@@ -34,6 +34,7 @@ public class FsignalrPlugin implements FlutterPlugin, Messages.HubConnectionMana
     public void onDetachedFromEngine(@NonNull FlutterPluginBinding binding) {
         Messages.HubConnectionManagerNativeApi.setUp(binding.getBinaryMessenger(), null);
         this.flutterPluginBinding = null;
+        clearResources();
     }
 
     @Override
@@ -49,11 +50,9 @@ public class FsignalrPlugin implements FlutterPlugin, Messages.HubConnectionMana
                     newHubConnectionManagerId
             );
             hubConnectionManagers.put(newHubConnectionManagerId, hubConnectionManager);
-            logger.info(
-                    "Created hub connection manager with id: "
-                            + newHubConnectionManagerId
-                            + ", managers count is now: "
-                            + hubConnectionManagers.size()
+            logger.info("Created hub connection manager with id: {}, managers count is now: {}",
+                    newHubConnectionManagerId,
+                    hubConnectionManagers.size()
             );
             final Messages.HubConnectionManagerIdMessage successResult =
                     new Messages.HubConnectionManagerIdMessage
@@ -68,7 +67,7 @@ public class FsignalrPlugin implements FlutterPlugin, Messages.HubConnectionMana
     }
 
     @Override
-    public void startHubConnection(@NonNull Messages.HubConnectionManagerIdMessage msg, @NonNull Messages.VoidResult result) {
+    public void startHubConnection(@NonNull Messages.HubConnectionManagerIdMessage msg, @NonNull Messages.Result<String> result) {
         final Long id = msg.getHubConnectionManagerId();
         HubConnectionManager hubConnectionManager = hubConnectionManagers.get(id);
         if (hubConnectionManager == null) {
@@ -89,6 +88,30 @@ public class FsignalrPlugin implements FlutterPlugin, Messages.HubConnectionMana
         }
 
         hubConnectionManager.stopHubConnection(result);
+    }
+
+    @Override
+    public void getConnectionId(@NonNull Messages.HubConnectionManagerIdMessage msg, @NonNull Messages.NullableResult<String> result) {
+        final Long id = msg.getHubConnectionManagerId();
+        HubConnectionManager hubConnectionManager = hubConnectionManagers.get(id);
+        if (hubConnectionManager == null) {
+            result.error(new Throwable(getHubConnectionManagerDoesNotExistMessage(id)));
+            return;
+        }
+
+        hubConnectionManager.getConnectionId(result);
+    }
+
+    @Override
+    public void getConnectionState(@NonNull Messages.HubConnectionManagerIdMessage msg, @NonNull Messages.Result<Messages.HubConnectionStateMessage> result) {
+        final Long id = msg.getHubConnectionManagerId();
+        HubConnectionManager hubConnectionManager = hubConnectionManagers.get(id);
+        if (hubConnectionManager == null) {
+            result.error(new Throwable(getHubConnectionManagerDoesNotExistMessage(id)));
+            return;
+        }
+
+        hubConnectionManager.getConnectionState(result);
     }
 
     @Override
@@ -126,5 +149,14 @@ public class FsignalrPlugin implements FlutterPlugin, Messages.HubConnectionMana
 
         hubConnectionManager.dispose(result);
         hubConnectionManagers.remove(id);
+    }
+
+    private void clearResources() {
+        for (var entry : hubConnectionManagers.entrySet()) {
+            entry.getValue().dispose();
+        }
+        hubConnectionManagers.clear();
+        nextHubConnectionManagerCreationId = 1;
+        logger.info("Disposed all hub connection managers");
     }
 }