+++ pigeons/fsignalr_pigeons.dart
@@ -129,11 +129,19 @@ abstract class HubConnectionManagerNativeApi {
 
   @async
   @TaskQueue(type: TaskQueueType.serialBackgroundThread)
-  void startHubConnection(HubConnectionManagerIdMessage msg);
+  String startHubConnection(HubConnectionManagerIdMessage msg);
 
   @async
   @TaskQueue(type: TaskQueueType.serialBackgroundThread)
   void stopHubConnection(HubConnectionManagerIdMessage msg);
+
+  @async
+  @TaskQueue(type: TaskQueueType.serialBackgroundThread)
+  String? getConnectionId(HubConnectionManagerIdMessage msg);
+
+  @async
+  @TaskQueue(type: TaskQueueType.serialBackgroundThread)
+  HubConnectionStateMessage getConnectionState(HubConnectionManagerIdMessage msg);
 
   @async
   @TaskQueue(type: TaskQueueType.serialBackgroundThread)