+++ CHANGELOG.md
@@ -6,3 +6,13 @@
 
 * Fixed some todos
 * Upgraded android gradle plugin version to 8.4.0
+
+## 1.2.0
+
+* Added new methods to the api to retrieve the connection id and connection state.
+* The `startConnection` method now returns the connection id.
+
+## 1.3.0
+
+* Updated dependencies
+* Clear resources in onDetachedFromEngine method